make it better

+++ database/migrations/2022_05_16_040748_create_products_table.php
@@ -19,7 +19,7 @@ class CreateProductsTable extends Migration
             $table->string('gambar_product');
             $table->string('jumlah_product');
             $table->text('deskripsi_product');
-            $table->double('harga');
+            $table->integer('harga');
 
             $table->timestamps();
         });
